% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracketsViewer.R
\name{bracketsViewer}
\alias{bracketsViewer}
\title{\if{html}{\out{
<Add Title>
}}}
\usage{
bracketsViewer(message, width = NULL, height = NULL, elementId = NULL)
}
\description{
\if{html}{\out{
<Add Description>
}}
}
