% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracketsViewer.R
\name{updateBracketsViewer}
\alias{updateBracketsViewer}
\title{Update bracketsViewer data}
\usage{
updateBracketsViewer(
  inputId,
  match_data,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{The input id that the data should be updated for}

\item{match_data}{The new data to use, a list containing the match data with the id}

\item{session}{The Shiny session object}
}
\description{
Update bracketsViewer data
}
